Move pages/404/styles.tsx to components/404/styles.tsx
