+++ components/404/styles.tsx
